now it matches the other directories
