+++ src/main/java/commandHandling/commands/placeCommands/PlaceGetFile.java
@@ -1,4 +1,4 @@
-package commandHandling.commands.place;
+package commandHandling.commands.placeCommands;
 
 import commandHandling.CommandContext;
 import commandHandling.CommandInterface;
