+++ src/main/java/commandHandling/commands/placeCommands/PlaceVerify.java
@@ -1,41 +1,45 @@
-package commandHandling.commands.place;
+package commandHandling.commands.placeCommands;
 
+import assets.CONFIG;
 import commandHandling.CommandContext;
 import commandHandling.CommandInterface;
 import net.dv8tion.jda.api.EmbedBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import services.database.DBHandlerConfig;
 import services.place.PlaceData;
 
 import java.util.List;
 
-public class PlaceStop implements CommandInterface {
-    private final Logger LOGGER = LoggerFactory.getLogger(PlaceStop.class);
+public class PlaceVerify implements CommandInterface {
+    private final Logger LOGGER = LoggerFactory.getLogger(PlaceVerify.class);
 
-    public PlaceStop(Logger cmdManagerLogger) {
+    public PlaceVerify(Logger cmdManagerLogger) {
         cmdManagerLogger.info("Loaded Command " + getName());
     }
 
     @Override
     public void handle(CommandContext ctx) {
-        PlaceData.stop = true;
+        PlaceData.verify = !PlaceData.verify;
+        DBHandlerConfig.updateConfig("PlaceVerify", "" + PlaceData.verify);
+        CONFIG.reload();
     }
 
     @Override
     public String getName() {
-        return "PlaceStop";
+        return "PlaceVerify";
     }
 
     @Override
     public EmbedBuilder getHelp() {
         EmbedBuilder embed = new EmbedBuilder();
-        embed.setDescription("Stops the drawing process");
+        embed.setDescription("Toggles whether or not placed pixels are getting verified");
         return embed;
     }
 
     @Override
     public List<String> getAliases() {
-        return List.of("ps");
+        return List.of("pv");
     }
 
     @Override
